Initial Upload
Odd hours on weekend to get the base in place.
Adding in menu to provide volume, score, timer and various levels of
difficulty on next pass.
Planning on changing revealed tiles to include a short animation of the
rock breaking into fragments.
Graphics are SGV - moving them into stylesheet soon and organizing for
quick theme changes.